minigbm: virtio_gpu: fill transfer offset
virglrenderer expects info->offset to be non-zero when a subregion
of the buffer is transferred.  Interestingly enough, minigbm doesn't
seem to hit that case in my (limited) testing.  Fill the transfer
offset and add logging statements.  Maybe we could fail on that case
if we don't encounter it in the wild.

BUG=none
TEST=manual testing with betty

Change-Id: I2f3afaff373777fb8808a0bea169be491d864d1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/2182791
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Chia-I Wu <olv@google.com>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>