minigbm: virtio_gpu: add BO_USE_HW_VIDEO_ENCODER usage to R8 format.
This combination will be used by ARCVM for video encoder bitstream
blobs. Minigbm uses the usage flags to determine whether it needs to
transfer data to/from the host, and whether that transfer needs to be
synchronous. BO_USE_HW_VIDEO_ENCODER needs to be set on the encoder's
output buffers so that minigbm knows that the host modifies the buffer,
so that minigbm transfers the data to the guest.

BUG=b:143570994
TEST=emerge-$BOARD minigbm arc-cros-gralloc

Change-Id: I05eb35d7b68092138c337c0e7fc43f60dd1bc32d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/2217833
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>