Fix amdgpu_drm/radeon_drm.h include paths
The paths work in the cros build environment, but not when building
on debian (which keeps them in libdrm/*).
The Makefile already takes care of adding the appropriate include
path, so we can just refer to the files directly.

Note: We do compile minigbm on debian for our Kokoro builder that
tests crosvm changes before the CQ.

BUG=None
TEST=CQ Dry Run

Change-Id: I766d14377941b2602df3c117055e17144af51b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/2643162
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Dennis Kempin <denniskempin@google.com>