minigbm: Android.bp: Add msm varients of gralloc.minigbm and libminigbm_cros_gralloc
In order to support msm devices, we need to enable -DDRV_MSM
as a build flag for both gralloc.minigbm and libminigbm_cros_gralloc.

This unfortunately requires new _msm post-fixed targets for both
gralloc.minigbm and libminigbm_cros_gralloc as well as for
the allocator and mapper HALs.

This feels a bit messy, but seems to be the only way to only
build with -DDRV_MSM on msm devices.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ic43967bddd40bab4622f92faebb96e679a19891d