minigbm: Restrict 256 alignment to grunt
This workaround addresses instability on grunt devices, but is a much
larger alignment than should be required. On zork, the larger alignment
ends up triggering a check that was added in mesa 21.1, causing an
allocation failure for some use cases.

Restrict the workaround to grunt devices, which are not affected by the
new check in mesa 21.1

BUG=b:180755705
TEST=start crostini on zork

Change-Id: I15855e06edb784fbca6b1128ed260450c2c82e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/2879313
Tested-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Bas Nieuwenhuizen <basni@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>