cros_gralloc: Check for init failure in allocator service
Prior to this change, the allocator service could be kept
alive in a bad state if the allocator failed to initialize
the underlying cros_gralloc_driver on the first attempt.

With this change, the allocator service will exit if the
allocator fails to initialize the underlying
cros_gralloc_driver which gives the service a chance to
restart and initialize again.

BUG=b:187802138
TEST=launch Cuttlefish

Change-Id: I2a0fd76cda8492632406597656a581325e69b18a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/3015656
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Tested-by: Jason Macnak <natsu@google.com>
Commit-Queue: Jason Macnak <natsu@google.com>