Add DRM_FORMAT_NV21 flex layout update in lockYCbCr
Some of the camera apps are requesting for NV21 streams.
As lockYCbCr fails with not supported, preview is not seen
with those apps.

Fix the issue by adding DRM_FORMAT_NV21 flex layoyt update
in

Note: Inorder to support apps requiring NV21 support, framework
and camera vHAL changes are required along with this change.

Change-Id: I625764149f1bd23298a9f4e58f756bd3a8adaa01
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>