[WA] Change linear alignment to 32 bytes
Vulkan will calculate row pitch itself and the alignment requirement is
different to minigbm. This will cause different layout view for EGL and
Vulkan. Now just change this linear alignment to 32 bytes to WA it. For
a formal fixing, we need to make the alignment are the same for both
side since framework will not pass stride information to Vulkan.

Tracked-On: OAM-101118
Signed-off-by: Marc Mao <marc.mao@intel.com>
Change-Id: Ibd4df67c5eacbc58013f6ea069662ee42ad376b8