gralloc: add workaround for MT8183 private format
Add support for the MT8183 private camera format to gralloc. This is
done as a vendor specific workaround that bypasses the normal minigbm
format resolution, targeting kukui+ARCVM.

This is aimed at working around the fact that the virtgpu cross domain
backend needs to directly resolve formats itself. Proper support for
involving the host when resolving formats would be cleaner. However,
that would be a substantial amount of work for what is (hopefully) a
one-off situation.

BUG=b:158617858
TEST=testBasicOpaqueToJpegReprocessing on kukui-arc-r DUT

Change-Id: I8e1d74aac0a16babd7d7e2e630ff8affb3104e8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/4258986
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>