[SMR0] Fixed the issue that sw codec fails to validate 10-bit type bu… …ffer
The function validateBufferSize fails when sw codec allocates buffer
with HAL_PIXEL_FORMAT_YCBCR_P010. The reason is that the pixel format
will be converted to drm format for allocating buffer, then
the format of the handle is assigned to HAL_PIXEL_ format type with
i915_private_invert_format, but the descriptor.format is still
the pixel format, they must be different when checked in
validateBufferSize. So we check the format type after uniformly
converting the pixel/HAL format to drm format.

Tracked-On:
Signed-off-by: zhangyichix <yichix.zhang@intel.com>