Fixed a performance issue for NeuSoft HMI app
Neusoft Demo app performance is low
It is caused by the Choreographer->doFrame->travesal->draw function.
This function will do the copyBlit operation.
The WB flag will use the cache to gain performance.
So we should not use the WC flag for the
case of BO_USE_SW_READ_OFTEN,due to the frequent CPU data reads.

Tracked-On:
Signed-off-by: Li,Gaoshun gaoshun.li@intel.com